requirements.txt
labscript-suite>=3.3.0
PyQt5>=5.15
numpy>=1.21,<2
scipy>=1.7
pyqtgraph>=0.12
h5py>=2.9
# Install wwlyn's PyRPL fork manually:
#   git clone -b max_hold_no_iir_improvement https://github.com/wwlyn/pyrpl_change.git
#   pip install ./pyrpl_change
